+++ src/main/java/com/wanted/teamr/tastyfinder/api/matzip/service/MatzipService.java
@@ -33,7 +33,7 @@ public class MatzipService {
     @Transactional(readOnly = true)
     public MatzipResponse getMatzip(Long matzipId) {
         Matzip matzip = getMatzipIfPresent(matzipId);
-        Long avgRating = calculateAvgRating(matzip);
+        double avgRating = calculateAvgRating(matzip);
         List<ReviewResponse> responseList = getReviewResponseList(matzip);
         return MatzipResponse.of(matzip, avgRating, responseList);
     }
@@ -43,8 +43,8 @@ public class MatzipService {
                 .orElseThrow(() -> new CustomException(ErrorCode.MATZIP_NOT_FOUND));
     }
 
-    public Long calculateAvgRating(Matzip matzip) {
-        return matzip.getTotalRating() / matzip.getReviewCount();
+    public double calculateAvgRating(Matzip matzip) {
+        return (double)(matzip.getTotalRating() / matzip.getReviewCount());
     }
 
     public List<ReviewResponse> getReviewResponseList(Matzip matzip) {
